+++ src/main/java/com/example/openschool1/repository/UserRepository.java
@@ -3,4 +3,4 @@ package com.example.openschool1.repository;
 import com.example.openschool1.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {}
+public interface UserRepository extends JpaRepository<User, Long> {}
